requirements.txt
twilio>=8.0.0
flask>=3.0.0
requests>=2.28.0
openai>=1.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
elevenlabs>=0.2.0
